# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Liquibase Project - OpenAPI 3.0
  description: |-
    API for managing Users, Products, and Orders
  contact:
    email: <a href="mailto:stefan.golubov@qinshift.com">stefan.golubov@qinshift.com</a>
  version: 1.0.0
...
paths:
...
  /products:
    put:
      tags:
        - products
      summary: Update an existing product
      description: Update an existing product by Id
      operationId: updateProduct
      requestBody:
        description: Update an existing product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Products'
          application/xml:
            schema:
              $ref: '#/components/schemas/Products'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Products'
        required: true
      responses:
        '200':
          description: Product has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/Products'
        '403':
          description: Access forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                type: Forbidden
                message: You don't have access to perform this operation
      security:
        - basicAuth: [ ]
    post:
      tags:
        - products
      summary: Add a new product
      description: Add a new product
      operationId: addProduct
      requestBody:
        description: Create a new product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProduct'
          application/xml:
            schema:
              $ref: '#/components/schemas/NewProduct'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewProduct'
        required: true
      responses:
        '200':
          description: Product has been successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Products'
            application/xml:
              schema:
                $ref: '#/components/schemas/Products'
        '403':
          description: Access forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                type: Forbidden
                message: You don't have access to perform this operation
      security:
        - basicAuth: [ ]
  /products/findByName:
    get:
      tags:
        - products
      summary: Find products by product name
      description: Multiple products can be provided with comma separated strings (case insensitive)
      operationId: findProductsByName
      parameters:
        - name: names
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Products have been successfully found by product name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Products'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Products'
      security:
        - basicAuth: [ ]
  /products/findAll:
    get:
      tags:
        - products
      summary: List all products
      operationId: findAllProducts
      responses:
        '200':
          description: Products have been successfully listed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Products'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Products'
      security:
        - basicAuth: [ ]
  /products/{id}:
    get:
      tags:
        - products
      summary: Find product by ID
      description: Returns a single product
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: Product ID to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product found by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Products'
            application/xml:
              schema:
                $ref: '#/components/schemas/Products'
      security:
        - basicAuth: [ ]
    post:
      tags:
        - products
      summary: Updates a product with form data
      description: ''
      operationId: updateProductWithForm
      parameters:
        - name: id
          in: path
          description: ID of product that needs to be updated
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: query
          description: Name for the product that needs to be updated
          schema:
            type: string
        - name: description
          in: query
          description: Description for the product that needs to be updated
          schema:
            type: string
        - name: price
          in: query
          description: Price for the product that needs to be updated
          schema:
            type: number
            format: double
        - name: quantity
          in: query
          description: Quantity of the product that needs to be updated
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Product found by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Access forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 403
                type: Forbidden
                message: You don't have access to perform this operation
      security:
        - basicAuth: [ ]
    delete:
      tags:
        - products
      summary: Deletes a product
      description: delete a product
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          description: Product ID for the product that needs to be deleted
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product found by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - basicAuth: [ ]
...
components:
  schemas:
...
    Products:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 6574214f-89e3-4001-b54d-0d185163f7a2
        name:
          type: string
          example: Smart TV
        description:
          type: string
          example: Samsung Smart TV
        price:
          type: string
          format: double
          example: 850.50
        quantity:
          type: integer
          format: int32
          example: 85
      xml:
        name: product
    NewProduct:
      type: object
      properties:
        name:
          type: string
          example: Smart TV
        description:
          type: string
          example: Samsung Smart TV
        price:
          type: string
          format: double
          example: 850.50
        quantity:
          type: integer
          format: int32
          example: 85
      xml:
        name: product
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic